requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
aiosqlite==0.20.0
aiosqlite3==0.3.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
frozenlist==1.4.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pydantic==2.8.2
pydantic_core==2.20.1
python-decouple==3.8
typing_extensions==4.12.2
yarl==1.13.1
